use a better name than "primitives"
