+++ Common/src/geometry/elements/CPRISM6.cpp
@@ -35,7 +35,7 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../../../include/geometry/primitives/CElement.hpp"
+#include "../../../include/geometry/elements/CElement.hpp"
 
 
 CPRISM6::CPRISM6() : CElementWithKnownSizes<NGAUSS,NNODE,NDIM>() {
