+++ Common/src/geometry/elements/CQUAD4.cpp
@@ -1,6 +1,6 @@
 /*!
- * \file CTRIA1.cpp
- * \brief Definition of the 3-node triangular element with one Gauss point.
+ * \file CQUAD4.cpp
+ * \brief Definition of the 4-node quadrilateral element with 4 Gauss points.
  * \author R. Sanchez
  * \version 6.2.0 "Falcon"
  *
@@ -35,18 +35,23 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../../../include/geometry/primitives/CElement.hpp"
+#include "../../../include/geometry/elements/CElement.hpp"
 
 
-CTRIA1::CTRIA1() : CElementWithKnownSizes<NGAUSS,NNODE,NDIM>() {
+CQUAD4::CQUAD4() : CElementWithKnownSizes<NGAUSS,NNODE,NDIM>() {
 
   /*--- Gauss coordinates and weights ---*/
 
-  GaussCoord[0][0] = 1.0/3.0;  GaussCoord[0][1] = 1.0/3.0;  GaussWeight(0) = 0.5;
+  su2double oneOnSqrt3 = 0.577350269189626;
+
+  GaussCoord[0][0] = -oneOnSqrt3;  GaussCoord[0][1] = -oneOnSqrt3;  GaussWeight(0) = 1.0;
+  GaussCoord[1][0] =  oneOnSqrt3;  GaussCoord[1][1] = -oneOnSqrt3;  GaussWeight(1) = 1.0;
+  GaussCoord[2][0] =  oneOnSqrt3;  GaussCoord[2][1] =  oneOnSqrt3;  GaussWeight(2) = 1.0;
+  GaussCoord[3][0] = -oneOnSqrt3;  GaussCoord[3][1] =  oneOnSqrt3;  GaussWeight(3) = 1.0;
 
   /*--- Store the values of the shape functions and their derivatives ---*/
 
-  unsigned short iGauss;
+  unsigned short iNode, iGauss;
   su2double Xi, Eta, val_Ni;
 
   for (iGauss = 0; iGauss < NGAUSS; iGauss++) {
@@ -54,35 +59,51 @@ CTRIA1::CTRIA1() : CElementWithKnownSizes<NGAUSS,NNODE,NDIM>() {
     Xi = GaussCoord[iGauss][0];
     Eta = GaussCoord[iGauss][1];
 
-    val_Ni = Xi;        GaussPoint[iGauss].SetNi(val_Ni,0);
-    val_Ni = Eta;       GaussPoint[iGauss].SetNi(val_Ni,1);
-    val_Ni = 1-Xi-Eta;  GaussPoint[iGauss].SetNi(val_Ni,2);
+    val_Ni = 0.25*(1.0-Xi)*(1.0-Eta);		GaussPoint[iGauss].SetNi(val_Ni,0);
+    val_Ni = 0.25*(1.0+Xi)*(1.0-Eta);		GaussPoint[iGauss].SetNi(val_Ni,1);
+    val_Ni = 0.25*(1.0+Xi)*(1.0+Eta);		GaussPoint[iGauss].SetNi(val_Ni,2);
+    val_Ni = 0.25*(1.0-Xi)*(1.0+Eta);		GaussPoint[iGauss].SetNi(val_Ni,3);
 
     /*--- dN/d xi, dN/d eta ---*/
 
-    dNiXj[iGauss][0][0] =  1.0;  dNiXj[iGauss][0][1] =  0.0;
-    dNiXj[iGauss][1][0] =  0.0;  dNiXj[iGauss][1][1] =  1.0;
-    dNiXj[iGauss][2][0] = -1.0;  dNiXj[iGauss][2][1] = -1.0;
+    dNiXj[iGauss][0][0] = -0.25*(1.0-Eta);  dNiXj[iGauss][0][1] = -0.25*(1.0-Xi);
+    dNiXj[iGauss][1][0] =  0.25*(1.0-Eta);  dNiXj[iGauss][1][1] = -0.25*(1.0+Xi);
+    dNiXj[iGauss][2][0] =  0.25*(1.0+Eta);  dNiXj[iGauss][2][1] =  0.25*(1.0+Xi);
+    dNiXj[iGauss][3][0] = -0.25*(1.0+Eta);  dNiXj[iGauss][3][1] =  0.25*(1.0-Xi);
 
   }
 
-  /*--- Shape functions evaluated at the nodes for extrapolation of the stresses at the Gaussian Points ---*/
-  /*--- The stress is constant over a TRIA1 element ---*/
+  /*--- Store the extrapolation functions (used to compute nodal stresses) ---*/
 
-  NodalExtrap[0][0] = 1.0;
-  NodalExtrap[1][0] = 1.0;
-  NodalExtrap[2][0] = 1.0;
+  su2double ExtrapCoord[4][2], sqrt3 = 1.732050807568877;;
+
+  ExtrapCoord[0][0] = -sqrt3;  ExtrapCoord[0][1] = -sqrt3;
+  ExtrapCoord[1][0] =  sqrt3;  ExtrapCoord[1][1] = -sqrt3;
+  ExtrapCoord[2][0] =  sqrt3;  ExtrapCoord[2][1] =  sqrt3;
+  ExtrapCoord[3][0] = -sqrt3;  ExtrapCoord[3][1] =  sqrt3;
+
+  for (iNode = 0; iNode < NNODE; iNode++) {
+
+    Xi = ExtrapCoord[iNode][0];
+    Eta = ExtrapCoord[iNode][1];
+
+    NodalExtrap[iNode][0] = 0.25*(1.0-Xi)*(1.0-Eta);
+    NodalExtrap[iNode][1] = 0.25*(1.0+Xi)*(1.0-Eta);
+    NodalExtrap[iNode][2] = 0.25*(1.0+Xi)*(1.0+Eta);
+    NodalExtrap[iNode][3] = 0.25*(1.0-Xi)*(1.0+Eta);
+
+  }
 
 }
 
-su2double CTRIA1::ComputeArea(const FrameType mode) const {
+su2double CQUAD4::ComputeArea(const FrameType mode) const {
 
   unsigned short iDim;
   su2double a[2] = {0.0,0.0}, b[2] = {0.0,0.0};
   su2double Area = 0.0;
 
   /*--- Select the appropriate source for the nodal coordinates depending on the frame requested
-        for the gradient computation, REFERENCE (undeformed) or CURRENT (deformed) ---*/
+        for the gradient computation, REFERENCE (undeformed) or CURRENT (deformed)---*/
   const su2activematrix& Coord = (mode==REFERENCE) ? RefCoord : CurrentCoord;
 
   for (iDim = 0; iDim < NDIM; iDim++) {
@@ -91,6 +112,13 @@ su2double CTRIA1::ComputeArea(const FrameType mode) const {
   }
 
   Area = 0.5*fabs(a[0]*b[1]-a[1]*b[0]);
+
+  for (iDim = 0; iDim < NDIM; iDim++) {
+    a[iDim] = Coord[0][iDim]-Coord[3][iDim];
+    b[iDim] = Coord[2][iDim]-Coord[3][iDim];
+  }
+
+  Area += 0.5*fabs(a[0]*b[1]-a[1]*b[0]);
 
   return Area;
