+++ Common/src/geometry/elements/CTRIA1.cpp
@@ -1,6 +1,6 @@
 /*!
- * \file CTETRA1.cpp
- * \brief Definition of 4-node tetra element with 1 Gauss point.
+ * \file CTRIA1.cpp
+ * \brief Definition of the 3-node triangular element with one Gauss point.
  * \author R. Sanchez
  * \version 6.2.0 "Falcon"
  *
@@ -35,73 +35,64 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../../../include/geometry/primitives/CElement.hpp"
+#include "../../../include/geometry/elements/CElement.hpp"
 
 
-CTETRA1::CTETRA1() : CElementWithKnownSizes<NGAUSS,NNODE,NDIM>() {
+CTRIA1::CTRIA1() : CElementWithKnownSizes<NGAUSS,NNODE,NDIM>() {
 
   /*--- Gauss coordinates and weights ---*/
 
-  GaussCoord[0][0] = 0.25;  GaussCoord[0][1] = 0.25; GaussCoord[0][2] = 0.25;  GaussWeight(0) = 1.0/6.0;
+  GaussCoord[0][0] = 1.0/3.0;  GaussCoord[0][1] = 1.0/3.0;  GaussWeight(0) = 0.5;
 
   /*--- Store the values of the shape functions and their derivatives ---*/
 
   unsigned short iGauss;
-  su2double Xi, Eta, Zeta, val_Ni;
+  su2double Xi, Eta, val_Ni;
 
   for (iGauss = 0; iGauss < NGAUSS; iGauss++) {
 
     Xi = GaussCoord[iGauss][0];
     Eta = GaussCoord[iGauss][1];
-    Zeta = GaussCoord[iGauss][2];
 
-    val_Ni = Xi;						  GaussPoint[iGauss].SetNi(val_Ni,0);
-    val_Ni = Eta;						  GaussPoint[iGauss].SetNi(val_Ni,1);
-    val_Ni = 1.0-Xi-Eta-Zeta;	GaussPoint[iGauss].SetNi(val_Ni,2);
-    val_Ni = Zeta;					  GaussPoint[iGauss].SetNi(val_Ni,3);
+    val_Ni = Xi;        GaussPoint[iGauss].SetNi(val_Ni,0);
+    val_Ni = Eta;       GaussPoint[iGauss].SetNi(val_Ni,1);
+    val_Ni = 1-Xi-Eta;  GaussPoint[iGauss].SetNi(val_Ni,2);
 
-    /*--- dN/d xi, dN/d eta, dN/d zeta ---*/
+    /*--- dN/d xi, dN/d eta ---*/
 
-    dNiXj[iGauss][0][0] =  1.0;  dNiXj[iGauss][0][1] =  0.0;  dNiXj[iGauss][0][2] =  0.0;
-    dNiXj[iGauss][1][0] =  0.0;  dNiXj[iGauss][1][1] =  1.0;  dNiXj[iGauss][1][2] =  0.0;
-    dNiXj[iGauss][2][0] = -1.0;  dNiXj[iGauss][2][1] = -1.0;  dNiXj[iGauss][2][2] = -1.0;
-    dNiXj[iGauss][3][0] =  0.0;  dNiXj[iGauss][3][1] =  0.0;  dNiXj[iGauss][3][2] =  1.0;
+    dNiXj[iGauss][0][0] =  1.0;  dNiXj[iGauss][0][1] =  0.0;
+    dNiXj[iGauss][1][0] =  0.0;  dNiXj[iGauss][1][1] =  1.0;
+    dNiXj[iGauss][2][0] = -1.0;  dNiXj[iGauss][2][1] = -1.0;
 
   }
 
   /*--- Shape functions evaluated at the nodes for extrapolation of the stresses at the Gaussian Points ---*/
-  /*--- The stress is constant at a TETRA1 element ---*/
+  /*--- The stress is constant over a TRIA1 element ---*/
 
   NodalExtrap[0][0] = 1.0;
   NodalExtrap[1][0] = 1.0;
   NodalExtrap[2][0] = 1.0;
-  NodalExtrap[3][0] = 1.0;
 
 }
 
-su2double CTETRA1::ComputeVolume(const FrameType mode) const {
+su2double CTRIA1::ComputeArea(const FrameType mode) const {
 
   unsigned short iDim;
-  su2double r1[3] = {0.0,0.0,0.0}, r2[3] = {0.0,0.0,0.0}, r3[3] = {0.0,0.0,0.0}, CrossProduct[3] = {0.0,0.0,0.0};
-  su2double Volume = 0.0;
+  su2double a[2] = {0.0,0.0}, b[2] = {0.0,0.0};
+  su2double Area = 0.0;
 
   /*--- Select the appropriate source for the nodal coordinates depending on the frame requested
-        for the gradient computation, REFERENCE (undeformed) or CURRENT (deformed)---*/
+        for the gradient computation, REFERENCE (undeformed) or CURRENT (deformed) ---*/
   const su2activematrix& Coord = (mode==REFERENCE) ? RefCoord : CurrentCoord;
 
   for (iDim = 0; iDim < NDIM; iDim++) {
-    r1[iDim] = Coord[1][iDim] - Coord[0][iDim];
-    r2[iDim] = Coord[2][iDim] - Coord[0][iDim];
-    r3[iDim] = Coord[3][iDim] - Coord[0][iDim];
+    a[iDim] = Coord[0][iDim]-Coord[2][iDim];
+    b[iDim] = Coord[1][iDim]-Coord[2][iDim];
   }
 
-  CrossProduct[0] = (r1[1]*r2[2] - r1[2]*r2[1])*r3[0];
-  CrossProduct[1] = (r1[2]*r2[0] - r1[0]*r2[2])*r3[1];
-  CrossProduct[2] = (r1[0]*r2[1] - r1[1]*r2[0])*r3[2];
+  Area = 0.5*fabs(a[0]*b[1]-a[1]*b[0]);
 
-  Volume = fabs(CrossProduct[0] + CrossProduct[1] + CrossProduct[2])/6.0;
-
-  return Volume;
+  return Area;
 
 }
